# Optimise checking the node types allowed for each rule