--- a/patches/nikic-php-parser-lib-phpparser-nodetraverser-php.patch
+++ b/patches/nikic-php-parser-lib-phpparser-nodetraverser-php.patch
@@ -47,7 +47,7 @@
 +    /**
 +     * @return NodeVisitor[]
 +     */
-+    protected function getVisitorsForNode(Node $node)
++    public function getVisitorsForNode(Node $node)
 +    {
 +        return $this->visitors;
      }