Modifies the PHP parser node traverser so that it can get a list of the visitors that apply to each node. When the visitors are Rector rules, we can compute which rules apply to each node class and cache this. By doing this we improve the performance of Rector by 20-25% according to my tests

To implement this optimisation we need to patch the PHP parser. In order to make this PR easy to review and test, I have added the patch to the code. If you agree that this code is fine I will create another PR in the `rectorphp/vendor-patches` repository with the patch and modify this PR to link to that patch instead of a local one