@TomasVotruba 2024-08-13
Thanks for proposal! This would be better to have directly in php-parser, as all projects would benefit.

Could you share `time` measurement on *your machine* with before/after results?