@carlos-granados 2024-08-13
A couple of time measurements in my machine:

- Running rector on the rector-src project. Before ~35s After ~27s
- Running rector on our company project. Before ~48s After ~35s

Regarding adding the code change to `php-parser`, I thought about this possibility but the fact is that for most cases this change would slow down the parser (not much) as it introduces an additional function call for each node traversed, so not sure it is appropriate. I can always create the PR and see what they think. And we can always add a patch ourselves and if the PR is accepted, remove the patch. What do you think?