@carlos-granados 2024-08-13
Timings:

Running Rector on the rector-src project:
Before `./bin/rector --clear-cache  220.79s user 12.11s system 657% cpu 35.440 total`
After `./bin/rector --clear-cache  168.53s user 11.80s system 638% cpu 28.236 total`

Running Rector on our company project:
Before `../rector-src/bin/rector --clear-cache  323.44s user 7.41s system 687% cpu 48.130 total`
After `../rector-src/bin/rector --clear-cache  236.92s user 6.64s system 678% cpu 35.903 total`

I am running this on a MacBook Pro with an M1 chip and 16gb of RAM. Php is 8.2.18