@samsonasik 2024-08-13
@carlos-granados I tested locally with only 8gb ram on m1 mac, and can see faster as well 👍 

**Before**

```
time bin/rector process --clear-cache
 2258/2258 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

 [OK] Rector is done!                                                                                                   

bin/rector process --clear-cache  234.69s user 11.00s system 711% cpu 34.540 total
```

**After**

```
time bin/rector process --clear-cache
 2258/2258 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

 [OK] Rector is done!                                                                                                   
                                                                                                                        
bin/rector process --clear-cache  173.57s user 10.82s system 709% cpu 25.982 total
```