@carlos-granados 2024-08-13
I added some tests for the `getVisitorsForNode()` function. I used the `PrivatesAccessor` tool to access the protected function, please let me know if this is OK