@carlos-granados 2024-08-16
@samsonasik thanks! Should I go ahead and create a PR in `rectorphp/vendor-patches` with the patch?