@carlos-granados 2024-08-17
I created the PR to add the patch to the `vendor-patches` repo https://github.com/rectorphp/vendor-patches/pull/8. Please let me know once it has been merged so that I can modify this PR to use that patch instead of the local one