@carlos-granados 2024-08-20
@TomasVotruba I updated the PR with the patch from `vendor-patches`