@TomasVotruba 2024-08-22
I thin this is ready to go.

Last but not least, I like to ask you @staabm if you have time. Could you run Blackfire before/after this PR? Just to see change in performnace/calls.