@staabm 2024-08-27
Started some testing but got interrupted. Will continue later today and report back