@staabm 2024-08-27
btw: since rector dropped the memory-peak information in the debug output I was not able to verify whether this PR introduces memory leak or consumes more ram