@TomasVotruba 2024-09-04
@staabm Thank you for confirmation :+1: 

Let's give this a try. We need to test downgrade itself + downgraded version before release, so there is a chance this might get reverted. 

Let's roll :)