@TomasVotruba 2024-09-04
@carlos-granados The version is now published. Could you try `rector/rector:dev-main` and measure time/memory before after?