@staabm 2024-09-04
btw: I checked the PHPStan codebase today whether there is a similar problem/opportunity.

I think the problem is solved differently in there
- https://github.com/phpstan/phpstan-src/blob/1.12.x/src/Rules/LazyRegistry.php

it seems to work in a way which does not need patching nikic/php-parser and does not rely on inheritance, which decouples the inner working of the php-parser internals.

maybe that something which can be taken as inspiration to solve this same problem in a more elegant way