@carlos-granados 2024-09-04
Before this improvement:
`vendor/bin/rector process --dry-run --clear-cache  488.44s user 12.92s system 540% cpu 1:32.76 total`
After this improvement:
`vendor/bin/rector process --dry-run --clear-cache  347.52s user 11.58s system 487% cpu 1:13.72 total`

This is when running Rector on my company's code base, about 1K files and about 100K lines of code