@carlos-granados 2024-09-04
> btw: I checked the PHPStan codebase today whether there is a similar problem/opportunity.
> 
> I think the problem is solved differently in there
> 
> * https://github.com/phpstan/phpstan-src/blob/1.12.x/src/Rules/LazyRegistry.php
> 
> it seems to work in a way which does not need patching nikic/php-parser and does not rely on inheritance, which decouples the inner working of PHPStan and php-parser internals.
> 
> maybe that something which can be taken as inspiration to solve this same problem in a more elegant way

@staabm PHPStan is able to use this because they don't use the node traverser from the php parser, they use their own traverser and in this way they are able to filter the rules that need to be applied to each node in a way which is similar to what I have added. I