@samsonasik 2024-09-04
I tested on a project with 4869 files:

**Before** v1.2.4

vendor/bin/rector process --clear-cache --dry-run  1629.64s user 49.34s system 651% cpu 4:17.87 total


**After** dev-main

vendor/bin/rector process --clear-cache --dry-run  1211.25s user 39.69s system 744% cpu 2:47.94 total