@TomasVotruba 2024-09-09
Numbers on one bigger project :partying_face: 

```bash
vendor/bin/rector -n --clear-cache  4735.58s user 111.60s system 983% cpu 8:13.08 total
```

After:

```bash
vendor/bin/rector -n --clear-cache  2887.89s user 73.34s system 1197% cpu 4:07.38 total
```