@samsonasik 2025-10-20
@TomasVotruba see PR on nikic/php-parser repo:

- https://github.com/nikic/PHP-Parser/pull/1122